Add empty commands to view and context menu
