+++ src/extension.ts
@@ -2,8 +2,8 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
+import { registerCommands } from "./workitems/commands";
 import { WorkItemTreeNodeProvider } from "./workitems/workitem.tree";
-import { GitExtension } from "./externals/git";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -12,15 +12,7 @@ export function activate(context: vscode.ExtensionContext) {
     treeDataProvider: new WorkItemTreeNodeProvider()
   });
 
-  vscode.commands.registerCommand("azure-boards.prefill", workItemId => {
-    const gitExtension = vscode.extensions.getExtension<GitExtension>(
-      "vscode.git"
-    );
-    if (gitExtension) {
-      const git = gitExtension.exports.getAPI(1);
-      git.repositories[0].inputBox.value = `Fix #${workItemId}`;
-    }
-  });
+  registerCommands();
 }
 
 // this method is called when your extension is deactivated
